requirements.txt
datasets>=2.21.0,<4.0.0
faiss-cpu>=1.8.0
matplotlib>=3.8.0
numpy>=1.26.0
PyYAML>=6.0.0
rank-bm25>=0.2.2
torch>=2.3.0
transformers>=4.44.0
tqdm>=4.66.0
pytest>=8.0.0
